+++ km_api/km_api/production_settings.py
@@ -61,6 +61,16 @@ DATABASES = {
 }
 
 
+# Authentication Configuration
+
+PASSWORD_RESET_EXPIRATION_HOURS = int(os.environ.get(
+    'PASSWORD_RESET_EXPIRATION_HOURS',
+    '1'))
+PASSWORD_RESET_LINK_TEMPLATE = os.environ.get(
+    'PASSWORD_RESET_LINK_TEMPLATE',
+    'https://example.com/change-password/?key={key}')
+
+
 # Email Settings
 
 AWS_SES_REGION_NAME = os.environ.get('AWS_REGION', 'us-east-1')